Move class SearchResource to rest package (was mistakenly put in wrong package). Method getDrives WIP.
